Move EmailOptions type to new types file in email folder
