+++ src/email/sendEmail/sendEmail.test.ts
@@ -3,7 +3,7 @@ import sendEmail from "./sendEmail.js";
 import transporter from "../transporter.js";
 import type SMTPTransport from "nodemailer/lib/smtp-transport/index.js";
 import smtpStatusCodes from "../../utils/smtpStatusCodes.js";
-import type { EmailOptions } from "../../types/types.js";
+import type { EmailOptions } from "../types/types.js";
 import { environment } from "../../loadEnvironments.js";
 
 const {
