+++ src/email/types/types.ts
@@ -1,0 +1,7 @@
+import type nodemailer from "nodemailer";
+import type { WithRequiredProperties } from "../../types/types";
+
+export type EmailOptions = WithRequiredProperties<
+  nodemailer.SendMailOptions,
+  "to" | "subject" | "text"
+>;
